+++ custom_components/metlink_explorer/__init__.py
@@ -5,18 +5,30 @@ import logging
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
+from .api import MetlinkApiClient
+from .coordinator import MetlinkDataUpdateCoordinator
+from .const import CONF_API_KEY, CONF_ROUTE_ID
 from .const import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
 
-PLATFORMS: list[Platform] = [Platform.SENSOR]
+PLATFORMS: list[Platform] = [Platform.SENSOR, Platform.SELECT]
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up Metlink Explorer from a config entry."""
     hass.data.setdefault(DOMAIN, {})
-    hass.data[DOMAIN][entry.entry_id] = {}
+    session = async_get_clientsession(hass)
+    api_client = MetlinkApiClient(entry.data[CONF_API_KEY], session)
+    coordinator = MetlinkDataUpdateCoordinator(hass, api_client, entry.data[CONF_ROUTE_ID])
+    await coordinator.async_config_entry_first_refresh()
+
+    hass.data[DOMAIN][entry.entry_id] = {
+        "api_client": api_client,
+        "coordinator": coordinator,
+    }
 
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
     return True